project/requirements.txt
flask
streamlit
gtts
nltk
pandas
textblob
beautifulsoup4
newspaper3k
requests
